Add files via upload
We will try to solve below problem using OpenAIGym environment.

Frozen Lake: Winter is here, You and your friends are tossing around the Frisbee at the park when you made a wild throw that left the Frisbee out in the middle of the lake. The water is mostly frozen, but there are a few holes where the ice has melted. If you stem in one of those holes, you will fall into the freezing water. T this time, there is an international Frisbee shortage, so it's absolutely imperative that navigate across the lake and retrieve the disk. However the ice is slippery so you won't always move in the direction you intend. The surface is described using a grid like the following...

S F F F

F H F H

F F F H

H F F G

Here S represents the starting point, it is safe. F Frozen surface which is also safe H is Hole, if is not good G is Goal

The agent can navigate Left, Right, Up and Down. It receives the reward as 1 if it reaches the goal, else 0. The episode ends when either the agent falls in Hole or reaches the Goal.